requirements.txt
aiosmtplib
alembic
anyio
asgiref
atomicwrites
attrs
bcrypt
blinker
certifi
cffi
charset-normalizer
click
colorama
cryptography
dnspython
ecdsa
email-validator
fastapi
fastapi-csv
fastapi-mail
FastAPI-SQLAlchemy
Flask
greenlet
h11
httptools
idna
iniconfig
is-disposable-email
itsdangerous
Jinja2
Mako
MarkupSafe
numpy
# orjson
packaging
pandas
passlib
Pillow
pluggy
py
pyasn1
pycparser
pydantic
PyJWT
pyotp
pyparsing
pytest
python-dateutil
python-dotenv
python-jose
python-multipart
pytz
PyYAML
qrcode
requests
rsa
six
sniffio
SQLAlchemy
starlette
tomli
typer
typing-extensions
# ujson
urllib3
uvicorn
watchgod
Werkzeug
